refactor polling sources
