+++ components/gmail/sources/new-attachment/new-attachment.mjs
@@ -1,5 +1,4 @@
 import common from "../common/polling.mjs";
-import constants from "../../common/constants.mjs";
 
 export default {
   ...common,
@@ -9,54 +8,8 @@ export default {
   version: "0.0.1",
   type: "source",
   dedupe: "unique",
-  props: {
-    ...common.props,
-    q: {
-      propDefinition: [
-        common.props.gmail,
-        "q",
-      ],
-    },
-    labels: {
-      propDefinition: [
-        common.props.gmail,
-        "label",
-      ],
-      type: "string[]",
-      label: "Labels",
-      optional: true,
-    },
-  },
-  hooks: {
-    async deploy() {
-      console.log(`Fetching last ${constants.HISTORICAL_EVENTS} historical events...`);
-      const response = await this.gmail.listMessages({
-        q: this.q,
-        labelIds: this.labels,
-        maxResults: constants.HISTORICAL_EVENTS,
-      });
-
-      const messageIds = response?.messages?.map((message) => message.id);
-      this.setLastMessageId(messageIds[0]);
-      const messages = this.filterMessagesWithAttachments(
-        await this.getMessages(messageIds.reverse()),
-      );
-      this.emitEvents(messages);
-    },
-  },
   methods: {
     ...common.methods,
-    async getMessages(ids = []) {
-      const promises = ids.map((id) => this.gmail.getMessage({
-        id,
-      }));
-      return Promise.all(promises);
-    },
-    filterMessagesWithAttachments(messages) {
-      return messages.filter(
-        (message) => message.payload.parts.filter((part) => part.body?.attachmentId).length,
-      );
-    },
     generateMeta(attachment, message) {
       return {
         id: attachment.body.attachmentId,
@@ -82,36 +35,16 @@ export default {
         }
       }
     },
-  },
-  async run() {
-    const lastMessageId = this.getLastMessageId();
-
-    const response = await this.gmail.listMessages({
-      q: this.q,
-      labelIds: this.labels,
-      maxResults: 100,
-    });
-
-    let messageIds = response?.messages?.map((message) => message.id);
-    this.setLastMessageId(messageIds[0]);
-    const index = messageIds.indexOf(lastMessageId);
-    if (index !== -1) {
-      messageIds = messageIds.slice(0, index);
-    }
-
-    const numMessages = messageIds.length;
-    if (!numMessages) {
-      console.log("No new message. Exiting...");
-      return;
-    }
-
-    const suffix = numMessages === 1
-      ? ""
-      : "s";
-    console.log(`Received ${numMessages} new message${suffix}. Please be patient while more information for each message is being fetched.`);
-    const messages = this.filterMessagesWithAttachments(
-      await this.getMessages(messageIds.reverse()),
-    );
-    this.emitEvents(messages);
+    filterMessagesWithAttachments(messages) {
+      return messages.filter(
+        (message) => message.payload.parts.filter((part) => part.body?.attachmentId).length,
+      );
+    },
+    async processMessageIds(messageIds) {
+      const messages = this.filterMessagesWithAttachments(
+        await this.getMessages(messageIds),
+      );
+      this.emitEvents(messages);
+    },
   },
 };
